Reworked donation page structure to allow for proper validation
